Improve locked UI and centralize lock text
